fix multiple rights changing bug
git-svn-id: https://waca.svn.sourceforge.net/svnroot/waca/trunk/acc@2228 41b7b2fa-e25d-4c8f-a56d-f93ce3f2ce94

+++ acc.php
@@ -798,6 +798,7 @@ elseif ($action == "usermgmt") {
 	if (isset ($_GET['demote'])) {
 		$did = sanitize($_GET['demote']);
 		$siuser = sanitize($_SESSION['user']);
+		if(hasright(getUsernameFromUid($did,"User"))) die("user is already demoted");
 		if (!isset($_POST['demotereason'])) {
 			echo "<h2>Demote Reason</h2><strong>The reason you enter here will be shown in the log. Please keep this in mind.</strong><br />\n<form action=\"acc.php?action=usermgmt&amp;demote=$did\" method=\"post\"><br />\n";
 			echo "<textarea name=\"demotereason\" rows=\"20\" cols=\"60\">";
@@ -836,6 +837,7 @@ elseif ($action == "usermgmt") {
 	if (isset ($_GET['suspend'])) {
 		$did = sanitize($_GET['suspend']);
 		$siuser = sanitize($_SESSION['user']);
+		if(hasright(getUsernameFromUid($did,"Suspended"))) die("user is already suspended");
 		if (!isset($_POST['suspendreason'])) {
 			echo "<h2>Suspend Reason</h2><strong>The user will be shown the reason you enter here. Please keep this in mind.</strong><br />\n<form action=\"acc.php?action=usermgmt&amp;suspend=$did\" method=\"post\"><br />\n";
 			echo "<textarea name=\"suspendreason\" rows=\"20\" cols=\"60\">";
@@ -872,6 +874,7 @@ elseif ($action == "usermgmt") {
 
 	}
 	if (isset ($_GET['promote'])) {
+		if(hasright(getUsernameFromUid($did,"Admin"))) die("user is already an admin");
 		$aid = sanitize($_GET['promote']);
 		$siuser = sanitize($_SESSION['user']);
 		$query = "UPDATE acc_user SET user_level = 'Admin' WHERE user_id = '$aid';";
